% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_sizes.R
\docType{data}
\name{body_sizes}
\alias{body_sizes}
\title{Bee body size measurments}
\format{A data frame with x rows and x variables}
\usage{
data(body_sizes)
}
\description{
A dataset containing bee body size for > 100 species.
}
\details{
\itemize{
  \item Family: Family of bee
  \item genus: Genus of bee  
  \item species: species of bee
  }
}
\keyword{datasets}
